Clean up the khangman.kcfg file, mostly eliminate the code="true"
According to the documentation:
An entry can optionally have a default value which is used as default when the
value isn't specified in any config file. Default values are interpreted as
literal constant values. If a default value needs to be computed or if it needs
to be obtained from a function call, the <default> tag should contain the
code="true" attribute. The contents of the <default> tag is then considered to
be a C++ expression.

Having removed that, the generated code looks alright wrt the default values. No
build and runtime regressions experienced on Archlinux.