Make the locally used variable local instead of a class member (cleanup)
REVIEW: 103912

There is apparently no user of the variable apart from right inside the same
method so the class can be cleaned up until the need arises for a class-scope
variable. The term "wordCount" is more consistent with "entryCount" and so on.
Thereby, if it is changed it can also be done better. Moreover, the class
variable lacked the usual m_* prefix, thus it was probably meant for local
usage.

Patch tested on Archlinux, and no regressions experienced.