Fix the "ends with new line" and pre/post increment general programming issue
1) Files that do not end with a newline character can cause problems. Please add
a newline character to the end of the file. It is mostly related for instance
very old gcc version, perhaps in the 3.x series, but stll: be precise.

2) Use ++ and -- as prefix whenever possible as these are more efficient than
postfix operators. Prefix increments first and then uses the variable, postfix
uses the actual; the variable is incremented as well. Because of this, the
prefix operators are inherently more efficient.

As a side effect: it apparently also fixes the "krazy" complains. Tested and no
regressions experienced.