Add an alphabet accessor method to the engine helper
It is simpler to do this way than generating the alphabet for the
listview/repeater's model in javascript. It is now being based upon the Ascii
table layout since the initial version is supposed to be English only. Further
improves and UTF8 character management will be added later on (definitely
planned).