Use mylist.at(i) instead of mylist.value(i) since it is faster in here
From the documentation:
"If the index i is out of bounds, the function returns a default-constructed
value. If you are certain that the index is going to be within bounds, you can
use at() instead, which is slightly faster."

We are certain in this special case it is within the bounds since it the
entryCount is just asked a little bit above.