Do not play the chalk sound once khangman-harmattan is launched
Unfortunately I need this initialized variable for the time being since it seems
to not be possible in QML to initialize an element property from the C++ side
for instance with an invokable method without hitting the onMyPropertyChanged
signal. As for that signal handler, there would be this really little sound
effect to be played when the user selects a different category, and a new word
is also generated.