Remove two unused compile options
Looking at the history, they are leftover from a time when it was
possible to compile 3 flavors: traditional application, Plasma Active
and Meego. Meego has been removed for a while, the Plasma Active
version removed during the KF5 porting, so there is "just" the
normal application. The variables are not required anymore.