Use theme letter color for labels on gui.
Before the KHangMan gui was changed to qml the themes used a letter
color and other colors to help keep contrast between the background
and the letters.
Expose KHangManTheme letter colors through KHangMan object as letterColor
property.
Use letterColor property for the labels in the gui that don't already
have a color tied to the meaning like wins does.