Fix KHangMan default language.
BUG: 392493
Somehow default language was set to "userLanguage" which was supposed
to be based on locale, but isn't when compiled by kcfg, it's a static
string, so just set the default back to en to be back to previous
behavior.
TODO: Actually use the user's locale to set the default language.