fix nextWord logic to prevent infinite loop in certain cases
If the word contains any of the revealed characters, such as dashes, spaces,
middot or apostrophe, the loop to update them in the visible word doesn't make
progress. It starts from the beginning of the original word each iteration.
Since there are no changes in the original word, if it loops, it loops forever.